python_worker/requirements.txt
celery[redis]==5.4.0
flask==3.0.3
redis==5.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
